feat: Fix release behavior
If we can push then we push tags and commits to GitHub either via HTTPS
(and GH_TOKEN) or SSH. If the user has a GH_TOKEN configured, we also
create a release on GitHub.
If the user has a CARGO_TOKEN configured, then we also release on
crates.io. This is happening independently from GitHub releases and
pushes